Rename `FoundationLMConversationThread` to `SystemLLMConversationThread`
