+++ Sources/SwiftAI/Backends/Apple/SystemLLM.swift
@@ -9,11 +9,9 @@ import Foundation
 ///
 /// ## Key Features
 ///
-/// - **On-device processing**: All inference happens locally, ensuring user privacy
-/// - **Tool calling**: Full support for function calling with automatic tool integration
-/// - **Structured output**: Generate typed Swift objects with comprehensive schema support
-/// - **Conversation threads**: Maintain stateful conversations with automatic context management
-/// - **Hybrid compatibility**: Seamlessly integrates with SwiftAI's model-agnostic API
+/// - **On-device processing**: All inference happens locally, ensuring user privacy at no cost.
+/// - **Tool calling**: Support for augmenting the LLM with custom functions to enhance its capabilities.
+/// - **Structured output**: Generate typed Swift objects.
 ///
 /// ## Requirements
 ///
@@ -98,7 +96,7 @@ import Foundation
 /// - Note: Always check `isAvailable` before making inference calls and handle errors appropriately.
 @available(iOS 26.0, macOS 26.0, *)
 public struct SystemLLM: LLM {
-  public typealias ConversationThread = FoundationLMConversationThread
+  public typealias ConversationThread = SystemLLMConversationThread
 
   private let model: SystemLanguageModel
 
@@ -135,8 +133,8 @@ public struct SystemLLM: LLM {
   public func makeConversationThread(
     tools: [any Tool],
     messages: [Message]
-  ) -> FoundationLMConversationThread {
-    return FoundationLMConversationThread(
+  ) -> SystemLLMConversationThread {
+    return SystemLLMConversationThread(
       model: model,
       tools: tools,
       messages: messages
@@ -186,7 +184,7 @@ public struct SystemLLM: LLM {
   public func reply<T: Generable>(
     to prompt: any PromptRepresentable,  // TODO: This should probably be a UserMessage to avoid `reply(to: AIMessage(...))`
     returning type: T.Type,
-    in thread: FoundationLMConversationThread,
+    in thread: SystemLLMConversationThread,
     options: LLMReplyOptions
   ) async throws -> LLMReply<T> {
     // TODO: Implement LLMReplyOptions support (temperature, maxTokens, etc.)
@@ -208,64 +206,5 @@ public struct SystemLLM: LLM {
       throw LLMError.generalError("Generation failed: \(error)")
     }
   }
-}
-
-/// A conversation thread that maintains stateful interactions with Apple's on-device language model.
-@available(iOS 26.0, macOS 26.0, *)
-public final actor FoundationLMConversationThread {
-  let session: LanguageModelSession
-
-  init(
-    model: SystemLanguageModel,
-    tools: [any Tool],
-    messages: [Message]
-  ) {
-    let transcript = FoundationModels.Transcript(messages: messages, tools: tools)
-    let foundationTools = tools.map { FoundationModelsToolAdapter(wrapping: $0) }
-
-    self.session = LanguageModelSession(
-      model: model,
-      tools: foundationTools,
-      transcript: transcript
-    )
-  }
-
-  func generateResponse<T: Generable>(
-    userMessage: Message,
-    type: T.Type
-  ) async throws -> LLMReply<T> {
-    let prompt = toFoundationPrompt(message: userMessage)
-
-    let content: T = try await {
-      if T.self == String.self {
-        let response: LanguageModelSession.Response<String> = try await session.respond(to: prompt)
-        return unsafeBitCast(response.content, to: T.self)
-      } else {
-        let response = try await session.respond(
-          to: prompt,
-          schema: try T.schema.toGenerationSchema()
-        )
-        // TODO: Add a protocol extension on `Generable` to conform `GeneratedContentConvertible`
-        // and use it here.
-        guard let jsonData = response.content.jsonString.data(using: .utf8) else {
-          throw LLMError.generalError("Failed to convert JSON string to Data")
-        }
-        return try JSONDecoder().decode(T.self, from: jsonData)
-      }
-    }()
-
-    let messages = try session.transcript.messages
-    return LLMReply(content: content, history: messages)
-  }
-
-  private func toFoundationPrompt(message: Message) -> FoundationModels.Prompt {
-    return FoundationModels.Prompt(message.text)
-  }
-}
-
-@available(iOS 26.0, macOS 26.0, *)
-private func mapAppleError(_ error: LanguageModelSession.GenerationError) -> LLMError {
-  // TODO: Implement proper error mapping when LLMError is fully defined
-  return .generalError("Apple generation error: \(error)")
 }
 #endif
